#38: ejected text to a separate package
- updated dependencies in related packages
- updated package.json to have valida author and contributor properties
- moved RandomString to text package as it's related to text manipulation scope

----
Refactoring: split core into individual packages
https://github.com/monumentjs/core/issues/38
