+++ packages/assert/package.json
@@ -10,10 +10,16 @@
   "license": "MIT",
   "homepage": "https://monumentjs.github.io/package/assert",
   "repository": "https://github.com/monumentjs/core.git",
+  "author": {
+    "name": "Alex Chugaev",
+    "email": "achugaev93@gmail.com",
+    "url": "https://github.com/avchugaev"
+  },
   "contributors": [
     {
-      "name": "Alex Chugaev",
-      "email": "achugaev93@gmail.com"
+      "name": "Dima Kolodko",
+      "email": "dima@kolodko.pro",
+      "url": "https://github.com/prowwid"
     }
   ],
   "dependencies": {
