+++ packages/core/package.json
@@ -12,15 +12,22 @@
   "license": "MIT",
   "homepage": "https://monumentjs.github.io/package/core",
   "repository": "https://github.com/monumentjs/core.git",
+  "author": {
+    "name": "Alex Chugaev",
+    "email": "achugaev93@gmail.com",
+    "url": "https://github.com/avchugaev"
+  },
   "contributors": [
     {
-      "name": "Alex Chugaev",
-      "email": "achugaev93@gmail.com"
+      "name": "Dima Kolodko",
+      "email": "dima@kolodko.pro",
+      "url": "https://github.com/prowwid"
     }
   ],
   "dependencies": {
     "reflect-metadata": "^0.1.13",
-    "@monument/collections": "^0.15.0"
+    "@monument/collections": "^0.15.0",
+    "@monument/text": "^0.15.0"
   },
   "publishConfig": {
     "access": "public"
