+++ packages/date/package.json
@@ -10,17 +10,25 @@
   "license": "MIT",
   "homepage": "https://monumentjs.github.io/package/date",
   "repository": "https://github.com/monumentjs/core.git",
+  "author": {
+    "name": "Alex Chugaev",
+    "email": "achugaev93@gmail.com",
+    "url": "https://github.com/avchugaev"
+  },
   "contributors": [
     {
-      "name": "Alex Chugaev",
-      "email": "achugaev93@gmail.com"
+      "name": "Dima Kolodko",
+      "email": "dima@kolodko.pro",
+      "url": "https://github.com/prowwid"
     }
   ],
   "publishConfig": {
     "access": "public"
   },
   "dependencies": {
-    "@monument/core": "^0.15.0"
+    "@monument/core": "^0.15.0",
+    "@monument/collections": "^0.15.0",
+    "@monument/text": "^0.15.0"
   },
   "gitHead": "9a16a9bdb55875bbad746859180800f7859832b6"
 }
