+++ packages/text/.npmignore
@@ -1,0 +1,5 @@
+node_modules
+test
+specs
+*.ts
+!*.d.ts
