+++ packages/text/README.md
@@ -1,0 +1,38 @@
+# Collections
+
+[![Travis CI Status](https://img.shields.io/travis/monumentjs/core/master.svg?logo=travis)](https://travis-ci.org/monumentjs/core)
+[![Slack Chat](https://img.shields.io/badge/slack-chat-brightgreen.svg?logo=slack)](https://join.slack.com/t/monumentjs/shared_invite/enQtNDY1ODA1MTExMzQ4LTI0MjllODEwOTk5MjM0NGIwY2YwNzVjNDU3YjEwYzYwYTNjMmI0NjFkNmNjMDFlMjA1NzgzODk0NjcxZTc4NjM)
+[![npm version](https://badge.fury.io/js/%40monument%2Ftext.svg)](https://badge.fury.io/js/%40monument%2Ftext)
+![npm version](https://david-dm.org/monumentjs/text.svg)
+
+Provides text manipulation solutions
+
+Related links:
+
+- [API reference](https://monumentjs.github.io/package/collections)
+- [Core API reference](https://monumentjs.github.io/package/core)
+
+# Install
+
+Install with NPM:
+
+```
+npm i @monument/text
+```
+
+Install with Yarn:
+
+```
+yarn add @monument/text
+```
+
+## Contribution
+
+Feel free to make pull-requests.
+Your help is always welcome.
+
+
+## License
+
+MIT
+
