+++ packages/text/index.ts
@@ -1,0 +1,11 @@
+export * from './src/Lines';
+export * from './src/StringBuilder';
+export * from './src/Strings';
+export * from './src/StringUtils';
+export * from './src/TemplateString';
+export * from './src/RandomString';
+export * from './src/parser/Parser';
+export * from './src/parser/ParsingException';
+export * from './src/parser/BooleanParser';
+export * from './src/parser/FloatParser';
+export * from './src/parser/IntParser';
