+++ packages/text/package.json
@@ -1,8 +1,17 @@
 {
-  "name": "@monument/keyboard",
+  "name": "@monument/text",
   "version": "0.15.0",
-  "description": "Template",
-  "keywords": [],
+  "description": "TypeScript text manipulation library",
+  "keywords": [
+    "typescript",
+    "text",
+    "standard",
+    "library",
+    "core"
+  ],
+  "license": "MIT",
+  "homepage": "https://monumentjs.github.io/package/text",
+  "repository": "https://github.com/monumentjs/core.git",
   "author": {
     "name": "Alex Chugaev",
     "email": "achugaev93@gmail.com",
@@ -16,7 +25,8 @@
     }
   ],
   "dependencies": {
-    "@monument/core": "^0.15.0"
+    "@monument/core": "^0.15.0",
+    "@monument/collections": "^0.15.0"
   },
   "publishConfig": {
     "access": "public"
