+++ packages/text/src/CappedString.ts
