+++ packages/text/src/RegExpUtils.ts
