+++ packages/uri/package.json
@@ -15,14 +15,22 @@
   "license": "MIT",
   "homepage": "https://monumentjs.github.io/package/uri",
   "repository": "https://github.com/monumentjs/core.git",
+  "author": {
+    "name": "Alex Chugaev",
+    "email": "achugaev93@gmail.com",
+    "url": "https://github.com/avchugaev"
+  },
   "contributors": [
     {
-      "name": "Alex Chugaev",
-      "email": "achugaev93@gmail.com"
+      "name": "Dima Kolodko",
+      "email": "dima@kolodko.pro",
+      "url": "https://github.com/prowwid"
     }
   ],
   "dependencies": {
-    "@monument/core": "^0.15.0"
+    "@monument/core": "^0.15.0",
+    "@monument/collections": "^0.15.0",
+    "@monument/text": "^0.15.0"
   },
   "publishConfig": {
     "access": "public"
