+++ packages/uri/src/QueryParameters.ts
@@ -1,4 +1,6 @@
-import { BooleanParser, FloatParser, IntParser, LinkedMultiValueMap, PreserveCaseEquals, ToString } from '@monument/core';
+import { PreserveCaseEquals, ToString } from '@monument/core';
+import { LinkedMultiValueMap } from '@monument/collections';
+import { BooleanParser, FloatParser, IntParser } from '@monument/text';
 import { ReadOnlyQueryParameters } from './ReadOnlyQueryParameters';
 import { QueryParameterValueEquals } from './QueryParameterValueEquals';
 import { UriConstants } from './UriConstants';
